+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+<a name="0.3.0-beta.2"></a>
+# 0.3.0-beta.2 (2018-01-24)
+
+
+### Bug Fixes
+
+* Print array types in usage ([04b04b9](https://github.com/ls-age/expose/commits/04b04b9))
+
+
+### Features
+
+* Add option for version and help ([4fa0c04](https://github.com/ls-age/expose/commits/4fa0c04))
+* Support array options ([#5](https://github.com/ls-age/expose/issues/5)) ([6809432](https://github.com/ls-age/expose/commits/6809432))
+
+
+
+
 <a name="0.2.0"></a>
 # 0.2.0 (2017-12-18)